Nmap scan rule groundwork
Basics of looking for an NMAP scan.

Records source IPs as a key in a dictionary, and stores dport and timestamps for each packet in a list.

To do:

Ignore ephemeral dst ports, finally check for packets that are too rapid, possibly store the packet number instead of the string, and add support for UDP and ICMP scans